Add OrderLineItem model

+++ server/db/models/Order.js
@@ -18,7 +18,7 @@ const Order = db.define('order', {
     defaultValue: 'created'
   },
   shippingTax: {
-    type: Sequelize.FLOAT
+    type: Sequelize.INTEGER
   }
   // Subtotal will be calculated as a function of the line items, promotion, and shipping tax as an instance method
 })
